extracted models and erased repairdatabase functions
