+++ src/database/database.ts
@@ -1,4 +1,5 @@
 import * as SQLite from "expo-sqlite";
+import { Product, InventoryItem, InventoryHistory, List, ShoppingListItem } from "./models/models";
 
 let db: SQLite.SQLiteDatabase | null = null;
 
@@ -93,174 +94,6 @@ const getDb = (): SQLite.SQLiteDatabase => {
     );
   }
   return db;
-};
-
-export interface Product {
-  id: number;
-  name: string; // "Rice", "Coffee", "Shampoo"
-  categoryId?: number; // FK to Category
-  createdAt: string;
-  updatedAt: string;
-}
-
-export interface Category {
-  id: number;
-  name: string;
-}
-
-export interface InventoryItem {
-  id: number;
-  listId: number; // FK to Context
-  productId: number; // FK to Product (generic)
-  quantity: number; // Current quantity of the *generic Product* (e.g., 3 units of "Rice")
-  sortOrder: number;
-  notes?: string; // Notes about this inventory item
-  updatedAt: string; // When was this stock last updated?
-  createdAt: string;
-}
-
-export interface InventoryHistory {
-  id: number;
-  listId: number; // FK to List
-  productId: number; // FK to Product (generic)
-  date: string; // YYYY-MM-DD
-  quantity: number;
-  notes?: string; // Optional notes for the history entry
-  createdAt: string;
-}
-
-export interface List {
-  id: number;
-  name: string;
-  order: number;
-}
-
-export interface ShoppingListItem {
-  id: number;
-  listId: number; // FK to List
-  productId: number; // FK to Product (generic)
-  quantity: number; // Desired quantity of the *generic Product*
-  checked: boolean; // True if item has been put in cart/purchased
-  price: number;
-  sortOrder: number;
-  notes?: string;
-  createdAt: string;
-  updatedAt: string;
-}
-
-const expectedSchemas = {
-  products: [
-    { name: "id", type: "INTEGER PRIMARY KEY AUTOINCREMENT" },
-    { name: "name", type: "TEXT NOT NULL UNIQUE" },
-    { name: "categoryId", type: "INTEGER" },
-    { name: "createdAt", type: "TEXT NOT NULL" },
-    { name: "updatedAt", type: "TEXT NOT NULL" },
-  ],
-  categories: [
-    { name: "id", type: "INTEGER PRIMARY KEY AUTOINCREMENT" },
-    { name: "name", type: "TEXT NOT NULL UNIQUE" },
-  ],
-  inventory_items: [
-    { name: "id", type: "INTEGER PRIMARY KEY AUTOINCREMENT" },
-    { name: "listId", type: "INTEGER NOT NULL" },
-    { name: "productId", type: "INTEGER NOT NULL" },
-    { name: "quantity", type: "INTEGER NOT NULL" },
-    { name: "sortOrder", type: "INTEGER NOT NULL", default: 0 },
-    { name: "notes", type: "TEXT" },
-    { name: "updatedAt", type: "TEXT NOT NULL" },
-    { name: "createdAt", type: "TEXT NOT NULL" },
-  ],
-  inventory_history: [
-    { name: "id", type: "INTEGER PRIMARY KEY AUTOINCREMENT" },
-    { name: "listId", type: "INTEGER NOT NULL" },
-    { name: "productId", type: "INTEGER NOT NULL" },
-    { name: "date", type: "TEXT NOT NULL" },
-    { name: "quantity", type: "INTEGER NOT NULL" },
-    { name: "notes", type: "TEXT" },
-    { name: "createdAt", type: "TEXT NOT NULL" },
-    { name: "UNIQUE", type: "(productId, date)" },
-  ],
-  lists: [
-    { name: "id", type: "INTEGER PRIMARY KEY AUTOINCREMENT" },
-    { name: "name", type: "TEXT NOT NULL UNIQUE" },
-    { name: "order", type: "INTEGER NOT NULL", default: 0 },
-  ],
-  shopping_list_items: [
-    { name: "id", type: "INTEGER PRIMARY KEY AUTOINCREMENT" },
-    { name: "listId", type: "INTEGER NOT NULL" },
-    { name: "productId", type: "INTEGER NOT NULL" },
-    { name: "quantity", type: "INTEGER NOT NULL" },
-    { name: "checked", type: "INTEGER NOT NULL", default: 0 },
-    { name: "price", type: "REAL" },
-    { name: "sortOrder", type: "INTEGER NOT NULL", default: 0 },
-    { name: "notes", type: "TEXT" },
-    { name: "createdAt", type: "TEXT NOT NULL" },
-    { name: "updatedAt", type: "TEXT NOT NULL" },
-  ],
-};
-
-const getExistingColumns = (tableName: string): string[] => {
-  try {
-    const result = getDb().getAllSync(`PRAGMA table_info(${tableName});`) as {
-      name: string;
-    }[];
-    return result.map((columnInfo) => columnInfo.name);
-  } catch (error) {
-    console.error(`Error getting column info for ${tableName}:`, error);
-    return [];
-  }
-};
-
-const addMissingColumn = (
-  tableName: string,
-  columnName: string,
-  columnType: string,
-  defaultValue: any
-) => {
-  const defaultValueClause =
-    typeof defaultValue !== "undefined" ? `DEFAULT ${defaultValue}` : "";
-  const alterStatement = `ALTER TABLE ${tableName} ADD COLUMN \`${columnName}\` ${columnType.replace(
-    "NOT NULL",
-    ""
-  )}${defaultValueClause};`;
-  getDb().execSync(alterStatement);
-};
-
-const repairDatabaseSchema = (tableName: string) => {
-  console.log("Repairing database schema for table:", tableName);
-  const columns = expectedSchemas[tableName];
-
-  // Check if the table exists
-  let tableExists = false;
-  try {
-    getDb().execSync(`SELECT 1 FROM ${tableName} LIMIT 1;`);
-    tableExists = true;
-  } catch (e: any) {
-    if (e.message.includes("no such table")) {
-      tableExists = false;
-    } else {
-      throw e; // Re-throw any other errors
-    }
-  }
-
-  if (!tableExists) {
-    const createStatement = `CREATE TABLE IF NOT EXISTS ${tableName} (${columns
-      .map((col) => `${col.name} ${col.type}`)
-      .join(", ")});`;
-    getDb().execSync(createStatement);
-  } else {
-    const existingColumns = getExistingColumns(tableName);
-    columns.forEach((expectedColumn) => {
-      if (!existingColumns.includes(expectedColumn.name)) {
-        addMissingColumn(
-          tableName,
-          expectedColumn.name,
-          expectedColumn.type,
-          expectedColumn.default
-        );
-      }
-    });
-  }
 };
 
 export const addList = async (
